requirements.txt
fastapi
uvicorn
httpx
watchfilespytest>=7.0.0
